requirements.txt
requests == 2.32.5
paho-mqtt == 2.1.0
